+++ lib/provider/settings_provider.dart
@@ -1,4 +1,16 @@
 import 'package:flutter/foundation.dart';
 
+enum StartView { AllPasswords, Folders, Favorites }
+
 class SettingsProvider with ChangeNotifier {
+  StartView _startView = StartView.AllPasswords;
+
+  get startView {
+    return _startView;
+  }
+
+  set startView(StartView startView) {
+    _startView = startView;
+    notifyListeners();
+  }
 }