+++ lib/screens/settings_screen.dart
@@ -1,0 +1,23 @@
+import 'package:flutter/material.dart';
+
+import '../widgets/app_drawer.dart';
+
+class SettingsScreen extends StatefulWidget {
+  static const routeName = '/settings';
+
+  @override
+  _SettingsScreenState createState() => _SettingsScreenState();
+}
+
+class _SettingsScreenState extends State<SettingsScreen> {
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      appBar: AppBar(
+        title: Text('Settings'),
+      ),
+      drawer: AppDrawer(),
+      body: Center(child: Text('TODO')),
+    );
+  }
+}